requirement.txt
torch
transformers
faiss-cpu
numpy
tqdm
langchain
langchain_text_splitters
sentence-transformers